LPS-63095 Move portal-configurator-extender to portal-configuration and rename portal-configuration-extender
