+++ modules/portal-configuration/portal-configuration-extender/build.gradle
