+++ modules/portal-configuration/portal-configuration-extender/src/main/java/com/liferay/portal/configuration/extender/NamedConfigurationContent.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.configuration.extender;
+
+import java.io.InputStream;
+
+/**
+ * @author Carlos Sierra Andrés
+ */
+public interface NamedConfigurationContent {
+
+	public InputStream getInputStream();
+
+	public String getName();
+
+}
