+++ modules/portal-configuration/portal-configuration-extender/src/testIntegration/resources/arquillian.xml
