+++ modules/portal-configuration/portal-configuration-extender/src/testIntegration/resources/framework.properties
