+++ modules/portal-configuration/portal-configuration-extender/test-bnd.bnd
@@ -1,0 +1,12 @@
+Bundle-Name: Liferay Portal Configurator Extender
+Bundle-SymbolicName: com.liferay.portal.configuration.extender
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.configuration.extender
+Import-Package: !javax.*,*
+Include-Resource:\
+	classes,\
+	test-classes/integration,\
+	@lib/biz.aQute.bndlib.jar!/aQute/lib/filter/*,\
+	@lib/org.apache.felix.utils.jar,\
+	@lib/portal-service.jar,\
+	@test-lib/com.liferay.registry.api.jar
