--- modules/portal/portal-configurator-extender/bnd.bnd
@@ -1,10 +1,0 @@
-Bundle-Name: Liferay Portal Configurator Extender
-Bundle-SymbolicName: com.liferay.portal.configurator.extender
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.portal.configurator.extender
-Include-Resource:\
-	@lib/org.apache.felix.utils.jar!/org/apache/felix/utils/extender/*,\
-	@lib/org.apache.felix.utils.jar!/org/apache/felix/utils/log/*
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Configuration Framework
--include: ../../../marketplace/platform/bnd.bnd
