--- modules/portal/portal-configurator-extender/test-bnd.bnd
@@ -1,12 +1,0 @@
-Bundle-Name: Liferay Portal Configurator Extender
-Bundle-SymbolicName: com.liferay.portal.configurator.extender
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.portal.configurator.extender
-Import-Package: !javax.*,*
-Include-Resource:\
-	classes,\
-	test-classes/integration,\
-	@lib/biz.aQute.bndlib.jar!/aQute/lib/filter/*,\
-	@lib/org.apache.felix.utils.jar,\
-	@lib/portal-service.jar,\
-	@test-lib/com.liferay.registry.api.jar
